predictionStockService/requirements.txt
flask

pandas
numpy

scikit-learn

psycopg2-binary
mysql-connector-python

py-eureka-client